+++ projects/ngx-pwa/local-storage/src/lib/lib.service.spec.ts
@@ -6,8 +6,9 @@ import { LocalStorage } from './lib.service';
 import { IndexedDBDatabase } from './databases/indexeddb-database';
 import { LocalStorageDatabase } from './databases/localstorage-database';
 import { MockLocalDatabase } from './databases/mock-local-database';
-import { JSONSchema, JSONSchemaString } from './validation/json-schema';
+import { JSONSchema } from './validation/json-schema';
 import { JSONValidator } from './validation/json-validator';
+import { SCHEMA_STRING, SCHEMA_NUMBER, SCHEMA_BOOLEAN } from './validation/constants';
 
 function testGetItem<T>(type: 'primitive' | 'object', localStorageService: LocalStorage, value: T, done: DoneFn) {
 
@@ -321,12 +322,9 @@ function tests(localStorageService: LocalStorage) {
     const value = {
       unexpected: 'value'
     };
-    // TODO: delete cast when TS 3.2 issue is fixed
     const schema: JSONSchema = {
       properties: {
-        expected: {
-          type: 'string'
-        } as JSONSchemaString
+        expected: SCHEMA_STRING
       },
       required: ['expected']
     };
@@ -357,12 +355,9 @@ function tests(localStorageService: LocalStorage) {
     const value = {
       expected: 'value'
     };
-    // TODO: delete cast when TS 3.2 issue is fixed
     const schema: JSONSchema = {
       properties: {
-        expected: {
-          type: 'string'
-        } as JSONSchemaString
+        expected: SCHEMA_STRING
       },
       required: ['expected']
     };
@@ -389,12 +384,9 @@ function tests(localStorageService: LocalStorage) {
 
   it('should return the data if the data is null (no validation)', (done: DoneFn) => {
 
-    // TODO: delete cast when TS 3.2 issue is fixed
     const schema: JSONSchema = {
       properties: {
-        expected: {
-          type: 'string'
-        } as JSONSchemaString
+        expected: SCHEMA_STRING
       },
       required: ['expected']
     };
@@ -949,15 +941,15 @@ describe('LocalStorage with IndexedDB', () => {
   }
 
   const getTestValues: [any, JSONSchema][] = [
-    ['hello', { type: 'string' }],
-    ['', { type: 'string' }],
-    [0, { type: 'number' }],
-    [1, { type: 'number' }],
-    [true, { type: 'boolean' }],
-    [false, { type: 'boolean' }],
+    ['hello', SCHEMA_STRING],
+    ['', SCHEMA_STRING],
+    [0, SCHEMA_NUMBER],
+    [1, SCHEMA_NUMBER],
+    [true, SCHEMA_BOOLEAN],
+    [false, SCHEMA_BOOLEAN],
     // TODO: delete cast when TS 3.2 issue is fixed
-    [[1, 2, 3], { items: { type: 'number' } } as JSONSchema],
-    [{ test: 'value' }, { properties: { test: { type: 'string' } } } as JSONSchema],
+    [[1, 2, 3], { items: SCHEMA_NUMBER }],
+    [{ test: 'value' }, { properties: { test: SCHEMA_STRING } }],
     [null, { type: 'null' }],
     [undefined, { type: 'null' }],
   ];